split SwiftUI files
