+++ experimental/examples/chat-mpp/iosApp/iosApp/YetAnotherSwiftUIScreen.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct YetAnotherSwiftUIScreen: View {
+    var body: some View {
+        GradientTemplate(title: "SwiftUI") {
+            VStack {
+                Text("Yet another SwiftUI screen")
+            }
+        }
+    }
+}
